Fix collecting static files in production mode.
As we have a settings module, its __init__.py is loaded at the beginning. In there we include the dev settings which modify
STATICFILES_DIRS. In production mode we load the production settings. As they are in the settings module, the dev settings are
loaded first. This leads to STATICFILES_DIRS including a directory that only exists in dev mode, resulting in an error.
Now, STATICFILES_DIRS does not include the dev-value when loading the production settings.